Add support for reporting movements and deaths
Extend cts-list into cts, with sub-commands for listing animals and
reporting deaths, movements off and movements on.

Allow the DDTS credentials, CTWS credentials and holding number to be
specified in the environment as well as hard-coded in the script.

CTWS exposes an annoying asynchronous interface for reporting movements
but the results of the validation can usually be retrieved immediately
after submission. Combine the two operations into a single easy-to-use
synchronous command in our interface, although this may need to retry
and thus be slow to complete if the request queue becomes backlogged.